Label text, Name on Card to Cardholder Name
This replaces usage of the text "Name on Card" with the text "Cardholder
Name".

There is evidence Name on Card gets misunderstood (or mistranslated) to
be asking for the card brand name like Visa / Mastercard.

The new phrase should afford the correct notion that the cardholders name
as printed on the card is the correct name to use.

Alternative text options I also considered (in case you want a different
one):

 - Full Name on Card
 - Your Name on Card
 - Cardholder Full Name

An alternative approach is to place an error message or warning when
someone attempts a checkout with a card brandname instead of their name.

I wasn't 100% sure cardholder names are always present on the cards (e.g.
prepaid) and thought this was a reasonable tradeoff in place of
implementing checkout validation warnings (we only support errors.)